Corrige l'acces du microblaze en DDR, calcul l'adresse sur 64 bits au lieu de 32 pour. Le microblaze ne pouvais pas lire le First frame ID et MOI freme ID correctement dépassé 4GB.
git-svn-id: http://einstein/svn/firmware/FIR-00257-Storage/trunk@16845 3e98269d-ff51-e449-abca-491a6567d830

+++ src/Buffering/Drivers/BufferManager.c
@@ -466,7 +466,7 @@ static uint32_t BufferManager_GetFrameId(t_bufferManager *pBufferCtrl, uint32_t 
       readAddrLoc = pBufferCtrl->Mem0_base_addr;
 
    // Since sequences are divided in the 2 memories, sequence offset and image offset are divided by 2
-   readAddrLoc += (SequenceID * pBufferCtrl->nbImagePerSeq + ImageLocation) / 2 * FrameSizeInBytes + FrameIDHdrAddr;
+   readAddrLoc += (uint64_t)(SequenceID * pBufferCtrl->nbImagePerSeq + ImageLocation) / 2 * (uint64_t)FrameSizeInBytes + FrameIDHdrAddr;
 
    // Convert address location to GPIO + AXI4L_32
    AXI4L_addrVal = BufferManager_MemAddrGPIO_Handler(readAddrLoc);
